put doughnut around video
