+++ typing-trainer/src/Components/Face/Face.scss
@@ -23,35 +23,68 @@
     display: flex; 
 
     flex-direction: column;
-    position: absolute;
-    top: 300px;
-    width: 400px; 
-    height: 400px; 
+    // position: absolute;
+    // top: 300px;
+    // width: 400px; 
+    // height: 400px; 
     align-items: center; 
-    justify-content: space-between;
+    // justify-content: space-between;
     -webkit-mask-image: -webkit-radial-gradient(circle, white 100%, black 100%);
-    -webkit-border-radius: 50%; 
-    margin-top: 160px;
+    // -webkit-border-radius: 50%; 
+    // margin-top: 160px;
+    // background-color: aqua;
+    border-style: solid;
+    
 }
 
 .face-video {
     position: absolute;
+    top: 200px;
+    height: 300px;
+    width: 300px;
     object-fit: cover;
     border-radius: 50%;
     background-color: #666;
+    align-items: center;
+    justify-content: center;
 }
 
 .face-canvas-calm, .face-canvas-not-calm {
     position: absolute;
-    top: -25px;
-    width: 680px;
-    height: 550px;
-    background-color: rgba(6, 1, 103, 0.432);
+    top: 200px;
+    // width: 680px;
+    // height: 550px;
+    height: 300px;
+    width: 300px;
+    // background-color: rgba(6, 1, 103, 0.432);
+    align-items: center;
+    justify-content: center;
 }
 .face-canvas-not-calm {
-    background-color: rgba(147, 66, 0, 0.599);
+    // background-color: rgba(147, 66, 0, 0.599);
+}
+
+.bar-chart {
+    // position: absolute;
+    width: 500px;
+    height: 500px
 }
 
 .hidden{
     display: none;
 }
+
+video {
+    height: 300px;
+    width: 300px;
+    border-radius: 50%;
+}
+
+
+// .face-feedback {
+//     display: absolute;
+//     flex-direction: row;
+//     width: 400px;
+
+// }
+
