+++ typing-trainer/src/Components/Home/Timer.jsx
@@ -18,42 +18,42 @@ export default function Timer({
   speed,
   setSpeed,
 }) {
-  const labels = ["neutral", "happy", "disgusted", "sad", "angry", "surprised"];
+  // const labels = ["neutral", "happy", "disgusted", "sad", "angry", "surprised"];
 
-  const data = {
-    labels,
-    datasets: [
-      {
-        label: "Emotion Feedback",
-        data: [
-          (((emotionLog.neutral - 1) / timeElapsed || 0) * 100).toFixed(2),
-          ((emotionLog.happy / timeElapsed || 0) * 100).toFixed(2),
-          ((emotionLog.disgusted / timeElapsed || 0) * 100).toFixed(2),
-          ((emotionLog.sad / timeElapsed || 0) * 100).toFixed(2),
-          ((emotionLog.angry / timeElapsed || 0) * 100).toFixed(2),
-          ((emotionLog.surprised / timeElapsed || 0) * 100).toFixed(2),
-        ],
+  // const data = {
+  //   labels,
+  //   datasets: [
+  //     {
+  //       label: "Emotion Feedback",
+  //       data: [
+  //         (((emotionLog.neutral - 1) / timeElapsed || 0) * 100).toFixed(2),
+  //         ((emotionLog.happy / timeElapsed || 0) * 100).toFixed(2),
+  //         ((emotionLog.disgusted / timeElapsed || 0) * 100).toFixed(2),
+  //         ((emotionLog.sad / timeElapsed || 0) * 100).toFixed(2),
+  //         ((emotionLog.angry / timeElapsed || 0) * 100).toFixed(2),
+  //         ((emotionLog.surprised / timeElapsed || 0) * 100).toFixed(2),
+  //       ],
 
-        backgroundColor: [
-          "rgba(54, 162, 235, 0.2)",
-          "rgba(255, 99, 132, 0.2)",
-          "rgba(255, 206, 86, 0.2)",
-          "rgba(75, 192, 192, 0.2)",
-          "rgba(153, 102, 255, 0.2)",
-          "rgba(255, 159, 64, 0.2)",
-        ],
-        borderColor: [
-          "rgba(54, 162, 235, 1)",
-          "rgba(255, 99, 132, 1)",
-          "rgba(255, 206, 86, 1)",
-          "rgba(75, 192, 192, 1)",
-          "rgba(153, 102, 255, 1)",
-          "rgba(255, 159, 64, 1)",
-        ],
-        borderWidth: 1,
-      },
-    ],
-  };
+  //       backgroundColor: [
+  //         "rgba(54, 162, 235, 0.2)",
+  //         "rgba(255, 99, 132, 0.2)",
+  //         "rgba(255, 206, 86, 0.2)",
+  //         "rgba(75, 192, 192, 0.2)",
+  //         "rgba(153, 102, 255, 0.2)",
+  //         "rgba(255, 159, 64, 0.2)",
+  //       ],
+  //       borderColor: [
+  //         "rgba(54, 162, 235, 1)",
+  //         "rgba(255, 99, 132, 1)",
+  //         "rgba(255, 206, 86, 1)",
+  //         "rgba(75, 192, 192, 1)",
+  //         "rgba(153, 102, 255, 1)",
+  //         "rgba(255, 159, 64, 1)",
+  //       ],
+  //       borderWidth: 1,
+  //     },
+  //   ],
+  // };
 
   useEffect(() => {
     let id;
@@ -72,7 +72,7 @@ export default function Timer({
   const minutes = timeElapsed / 60;
 
   return (
-    <>
+    <div className="timerContainer">
       <Fragment>
         <div className="feedback">
           <div className="time-elapsed-and-speed">
@@ -86,14 +86,10 @@ export default function Timer({
                 : `We couldn't detect you ${undetected} times during the exercise.`}
             </p>
           </div>
-          <div className="face-feedback">
-            <div className="bar-chart">
-              <Doughnut data={data} />
-            </div>
-          </div>
+
           <br />
         </div>
       </Fragment>
-    </>
+    </div>
   );
 }
